Fix media posted in one cliques shows up in all.
Media tab of each clique was showing media from all cliques, instead of
showing the media of the current cliques only.
For example if a person has 3 cliques A,B and C , the media Posted in
any of the tab will show up in all A, B, C cliques.
This behaviour is fixed in this commit.